Move ./database to ./app/database
